+++ tpu_inference/layers/vllm/quantization/nvfp4.py
@@ -50,8 +50,8 @@ import tpu_inference.envs as envs
 from tpu_inference.layers.common.linear import sharded_quantized_matmul
 from tpu_inference.layers.common.moe import MoEBackend
 from tpu_inference.layers.common.process_weights.linear_weights import (
-    LinearWeights, format_linear_scale, process_linear_weights,
-    shard_linear_weights, to_parameter_list)
+    LinearWeights, process_linear_weights, shard_linear_weights,
+    to_parameter_list)
 from tpu_inference.layers.common.process_weights.moe_weights import (
     FusedMoEWeights, process_quantized_moe_weights, shard_moe_weights_to_tpu)
 from tpu_inference.layers.common.quant_methods import NVFP4
@@ -190,18 +190,14 @@ class VllmNvfp4LinearMethod(VllmUnquantizedLinearMethod):
             fused=self.linear_config.fuse_matmuls,
             output_sizes=self.linear_config.output_sizes,
             reorder_size=self.linear_config.n_shards,
-            transposed=False,
+            enable_kernel=self.linear_config.enable_quantized_matmul_kernel,
         )
-        weights.weight_scale = format_linear_scale(
-            weights.weight_scale,
-            self.linear_config.enable_quantized_matmul_kernel)
         weights = torch_view(
             shard_linear_weights(
                 weights,
                 mesh=self.linear_config.mesh,
                 weight_p_spec=self.linear_config.weight_sharding,
                 bias_p_spec=self.linear_config.bias_sharding,
-                transposed=False,
             ))
 
         if self.linear_config.fuse_matmuls:
